phase3-chatbot/backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
slowapi==0.1.9
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.11
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# AI & MCP Integration
openai==1.3.5
websockets==12.0

# Utilities
uuid==1.30
redis==5.0.1

# Development
ruff==0.1.14
black==24.1.1
